requirements.txt
python-dotenv==1.0.0
openai==1.4.0
deepgram-sdk==2.12.0
pyaudio==0.2.14
rhasspy-silence==0.4.0
elevenlabs==0.2.27
pygame==2.5.2
taipy==3.0.0
pyttsx3